requirements.txt
flask==3.0.3
boto3==1.34.144
mysql-connector-python==9.0.0
bcrypt==4.1.3
python-dotenv==1.0.1
flask-session==0.8.0
flask-bcrypt==1.0.1  # Flask wrapper for bcrypt password hashing